Resume/requirements.txt
blinker==1.6.2
jwt==1.3.1
pymongo==4.3.3
python-dotenv==1.0.0
protobuf==3.19
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.12.0
Flask==2.3.2
fsspec==2023.5.0
fuzzywuzzy==0.18.0
gensim==4.3.1
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
inexactsearch==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Levenshtein==0.21.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
PyPDF2==3.0.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-Levenshtein==0.21.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
soundex==1.1.3
# spellchecker==0.4
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.6
Werkzeug==2.3.4
zipp==3.15.0